Move GenerateRoleName into utils.go
