+++ controllers/iamrole_controller_test.go
@@ -1,81 +1,13 @@
 package controllers_test
 
 import (
-	"context"
-	"github.com/golang/mock/gomock"
 	iammanagerv1alpha1 "github.com/keikoproj/iam-manager/api/v1alpha1"
 	. "github.com/keikoproj/iam-manager/controllers"
 	. "github.com/keikoproj/iam-manager/internal/config"
 	. "github.com/onsi/ginkgo"
 	. "github.com/onsi/gomega"
-	"gopkg.in/check.v1"
-	v1 "k8s.io/api/core/v1"
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"sigs.k8s.io/controller-runtime/pkg/event"
-	"testing"
 )
-
-type ControllerSuite struct {
-	t        *testing.T
-	ctx      context.Context
-	mockCtrl *gomock.Controller
-}
-
-func TestControllerSuite(t *testing.T) {
-	check.Suite(&ControllerSuite{t: t})
-	check.TestingT(t)
-}
-
-func (s *ControllerSuite) SetUpTest(c *check.C) {
-	s.ctx = context.Background()
-	s.mockCtrl = gomock.NewController(s.t)
-}
-
-func (s *ControllerSuite) TearDownTest(c *check.C) {
-	s.mockCtrl.Finish()
-}
-
-func (s *ControllerSuite) TestGenerateNameFunction(c *check.C) {
-	cm := &v1.ConfigMap{
-		Data: map[string]string{
-			"iam.role.derive.from.namespace": "false",
-			"iam.role.prefix":                "pfx",
-			"iam.role.separator":             "+",
-		},
-	}
-	Props = nil
-	err := LoadProperties("", cm)
-	c.Assert(err, check.Equals, nil)
-
-	resource := &iammanagerv1alpha1.Iamrole{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      "foo",
-			Namespace: "test-ns",
-		},
-	}
-	c.Assert(GenerateRoleName(*resource, *Props), check.Equals, "pfx+foo")
-}
-
-func (s *ControllerSuite) TestGenerateNameFunctionWithDeriveFromNamespaceEnabled(c *check.C) {
-	cm := &v1.ConfigMap{
-		Data: map[string]string{
-			"aws.accountId":                  "012345678901",  // Required mock for testing
-			"iam.role.derive.from.namespace": "true",
-			"iam.role.prefix":                "pfx",
-			"iam.role.separator":             "+",
-		},
-	}
-	Props = nil
-	LoadProperties("", cm)
-
-	resource := &iammanagerv1alpha1.Iamrole{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      "foo",
-			Namespace: "test-ns",
-		},
-	}
-	c.Assert(GenerateRoleName(*resource, *Props), check.Equals, "pfx+test-ns")
-}
 
 var _ = Describe("IamroleController", func() {
 	Describe("When checking a StatusUpdatePredicate", func() {
